backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
pillow==11.3.0
pytesseract==0.3.13
gunicorn==23.0.0
